requirements.txt
aniso8601==9.0.1
click==8.0.1
decamelize==0.1.2
Flask==2.0.1
Flask-RESTful==0.3.9
greenlet==1.1.0
pyhumps==3.0.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pydantic==1.8.2
PyMySQL==1.0.2
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
typing-extensions==3.10.0.0
Werkzeug==2.0.1
